front-end is done

+++ back-end/FaleMaisDomain/src/Entities/Call.cs
@@ -4,7 +4,7 @@ namespace FaleMaisDomain.Entities {
   public class Call {
     public byte FromAreaCode { get; set; }
     public byte ToAreaCode { get; set; }
-    public double PricePerMinute { get; set; }
+    public decimal PricePerMinute { get; set; }
     public DateTime? CreatedAt { get; set; }
     public DateTime? UpdatedAt { get; set; }
     public DateTime? DeletedAt { get; set; }
